Extracts the error to separate file
